requirements.txt
-e .
django-templated-mail==1.0.0
djangorestframework-jwt==1.11.0
djet==0.2.1
drf-tracking==1.3.1
social-auth-app-django==2.0.0
coverage==4.3.1

pytest==3.1.3
pytest-cov==2.5.1
pytest-django==3.1.2
flake8==3.4.1